requirements.txt
google-generativeai
docling
sentence-transformers
faiss-cpu
python-dotenv
transformers